+++ packages/controls/js/libs/border-radius-control/test/border-radius-control.cy.js
@@ -73,13 +73,20 @@ describe('border-radius-control component testing', () => {
 					name,
 				});
 
-				cy.get('input').clear();
-				cy.get('input').type(10);
+				cy.getByDataTest('border-radius-input-all').clear({
+					force: true,
+				});
+				cy.getByDataTest('border-radius-input-all').type('10', {
+					force: true,
+				});
 
-				cy.get('input').should('have.value', '10');
+				cy.getByDataTest('border-radius-input-all').should(
+					'have.value',
+					'10'
+				);
 
 				//Check data provider value
-				cy.get('input').then(() => {
+				cy.getByDataTest('border-radius-input-all').then(() => {
 					expect('10px').to.be.equal(getControlValue(name).all);
 				});
 			});
@@ -420,7 +427,9 @@ describe('border-radius-control component testing', () => {
 			name,
 		});
 
-		cy.get('input').type('5');
+		cy.getByDataTest('border-radius-input-all').type('5', {
+			force: true,
+		});
 
 		cy.get('@onChange').should('have.been.called');
 	});