+++ packages/controls/js/libs/box-position-control/test/box-position-control.cy.js
@@ -10,6 +10,20 @@ import BoxPositionControl from '../index';
 import { getControlValue } from '../../../store/selectors';
 import { absolutePositionValues } from './data/absolute-position-values';
 
+/**
+ * CustomSelectControl (@wordpress/components / Ariakit) renders a listbox with
+ * [role="option"] items, not ul/li.
+ */
+const selectBoxPositionType = (label) => {
+	cy.get('[data-cy="box-position-control"]')
+		.find('[aria-haspopup="listbox"]')
+		.click();
+	cy.get('[role="listbox"]:visible')
+		.find('[role="option"]')
+		.contains(label)
+		.click();
+};
+
 describe('box position control component testing', () => {
 	beforeEach(() => {
 		// run these tests as if in a desktop
@@ -47,8 +61,7 @@ describe('box position control component testing', () => {
 				name,
 			});
 
-			cy.get('.blockera-control-select').click();
-			cy.get('ul > li').contains('Relative').click();
+			selectBoxPositionType('Relative');
 
 			// Check data provider value!
 			cy.then(() => {
@@ -69,8 +82,7 @@ describe('box position control component testing', () => {
 				name,
 			});
 
-			cy.get('.blockera-control-select').click();
-			cy.get('ul > li').contains('Relative').click();
+			selectBoxPositionType('Relative');
 			cy.get('@onChangeMock').should('have.been.called');
 
 			// Check data provider value!
@@ -102,8 +114,7 @@ describe('box position control component testing', () => {
 				name,
 			});
 
-			cy.get('.blockera-control-select').click();
-			cy.get('ul > li').contains('Relative').click();
+			selectBoxPositionType('Relative');
 
 			// Check data provider value!
 			cy.then(() => {
@@ -140,8 +151,7 @@ describe('box position control component testing', () => {
 				),
 				name,
 			});
-			cy.get('.blockera-control-select').click();
-			cy.get('ul > li').contains('Relative').click();
+			selectBoxPositionType('Relative');
 
 			cy.get(
 				'span[aria-label="Top Position"][data-cy="label-control"]'
@@ -231,8 +241,7 @@ describe('box position control component testing', () => {
 				),
 				name,
 			});
-			cy.get('.blockera-control-select').click();
-			cy.get('ul > li').contains('Relative').click();
+			selectBoxPositionType('Relative');
 
 			const positions = [
 				'Top Position',
@@ -285,8 +294,7 @@ describe('box position control component testing', () => {
 				name,
 			});
 
-			cy.get('.blockera-control-select').click();
-			cy.get('ul > li').contains('Relative').click();
+			selectBoxPositionType('Relative');
 
 			// add top position
 			cy.get(
@@ -368,8 +376,7 @@ describe('box position control component testing', () => {
 				name,
 			});
 
-			cy.get('.blockera-control-select').click();
-			cy.get('ul > li').contains('Absolute').click();
+			selectBoxPositionType('Absolute');
 
 			// check absolute top left
 			cy.getByDataCy('absolute-top-left').click();
@@ -471,8 +478,7 @@ describe('box position control component testing', () => {
 				name,
 			});
 
-			cy.get('.blockera-control-select').click();
-			cy.get('ul > li').contains('Sticky').click();
+			selectBoxPositionType('Sticky');
 
 			// check sticky to top
 			cy.getByDataCy('stick-to-top').click();