+++ packages/editor/js/extensions/libs/typography/compatibility/letter-spacing.js
@@ -4,6 +4,7 @@
  * Blockera dependencies
  */
 import { isSpecialUnit } from '@blockera/controls';
+import { normalizeCssLengthValue } from '@blockera/utils';
 
 /**
  * Internal dependencies
@@ -32,7 +33,7 @@ export function letterSpacingFromWPCompatibility({
 		letterSpacing !== undefined
 	) {
 		attributes.blockeraLetterSpacing = {
-			value: letterSpacing,
+			value: normalizeCssLengthValue(letterSpacing),
 		};
 	}